requirements.txt
Django == 1.11.7
djangorestframework == 3.6.2